Added Stream.cpp and fixed use of _cplusplus
The implementation of the Stream class is missing. I added only
those missing parts that are needed to copile an executable based
on Kaleidoscope-Hardware-Virtual. Without this, the symbol would
be reported as undefined during linking.

The _cplusplus in stdlib_ext.h was misspelled it must actualy read "__cplusplus"
to work.